refactor: access token ecoded된 상태로 app state 저장
